fix: update the benchmark package
