+++ classifier/__init__.py
@@ -1,0 +1,43 @@
+from ._adaptive_random_forest import AdaptiveRandomForestClassifier
+from ._efdt import EFDT
+from ._hoeffding_tree import HoeffdingTree
+from ._naive_bayes import NaiveBayes
+from ._online_bagging import OnlineBagging
+from ._online_adwin_bagging import OnlineAdwinBagging
+from ._leveraging_bagging import LeveragingBagging
+from ._passive_aggressive_classifier import PassiveAggressiveClassifier
+from ._sgd_classifier import SGDClassifier
+from ._knn import KNN
+from ._sgbt import StreamingGradientBoostedTrees
+from ._oza_boost import OzaBoost
+from ._majority_class import MajorityClass
+from ._no_change import NoChange
+from ._online_smooth_boost import OnlineSmoothBoost
+from ._srp import StreamingRandomPatches
+from ._hoeffding_adaptive_tree import HoeffdingAdaptiveTree
+from ._samknn import SAMkNN
+from ._dynamic_weighted_majority import DynamicWeightedMajority
+from ._csmote import CSMOTE
+
+__all__ = [
+    "AdaptiveRandomForestClassifier",
+    "EFDT",
+    "HoeffdingTree",
+    "NaiveBayes",
+    "OnlineBagging",
+    "OnlineAdwinBagging",
+    "LeveragingBagging",
+    "KNN",
+    "PassiveAggressiveClassifier",
+    "SGDClassifier",
+    "StreamingGradientBoostedTrees",
+    "OzaBoost",
+    "MajorityClass",
+    "NoChange",
+    "OnlineSmoothBoost",
+    "StreamingRandomPatches",
+    "HoeffdingAdaptiveTree",
+    "SAMkNN",
+    "DynamicWeightedMajority",
+    "CSMOTE"
+]
